// src/app/shared/models/table/enum/table-column-filter-type.enum.ts
export const enum TableColumnFilterType {
    Contains = 'contains',
    StartsWith = 'startsWith',
    EndsWith = 'endsWith',
    Equal = 'equal',
    GreaterThan = 'greaterThan',
    GreaterThanOrEqual = 'greaterThanOrEqual',
    LessThan = 'lessThan',
    LessThanOrEqual = 'lessThanOrEqual',
    MultipleChoices = 'MultipleChoices',
}